inline issue (sloved)
