+++ SchedulingSimulator/cpu_scheduler.h
@@ -1,15 +1,18 @@
-#pragma once
 #include "waiting_queue.h"
+#ifndef _CPU_SCHEDULER_H
+#define _CPU_SCHEDULER_H
 
 
 class CpuScheduler
 {
 public:
-	CpuScheduler() : time(0), isRunning(false), isPreemptive(false), isRoundRobin(false) {}
+	CpuScheduler() : time(0), isRunning(false), isPreemptive(false), isRoundRobin(false), timeQuantum(-1) {}
 	void SetProcessQueue(std::unique_ptr<ProcessQueue> pQ)
 		{ this->pQ = std::move(pQ); }
 	void SetAlgorithm(Scheduling wQalgorithm, bool isPreemptive, bool isRoundRobin)
 		{ wQ.SetAlgorithm(wQalgorithm); this->isPreemptive = isPreemptive; this->isRoundRobin = isRoundRobin; }
+	void SetTimeQuantum(const double tq)
+		{ timeQuantum = tq;}
 
 	const waitingQueue& GetWQhandler() const
 		{ return wQ; }
@@ -29,88 +32,11 @@ private:
 	GanttChart ganttChart;
 	Process currentProcess;
 	double time;
+	double timeQuantum;
 
 	bool isPreemptive;
 	bool isRoundRobin;
 	bool isRunning;
 };
 
-
-
-void CpuScheduler::StepForward()
-{
-	// CPU scheduler 초기화
-	if (!isRunning) {
-
-		if (!pQ)
-			return;
-		wQ.Clear();
-		ganttChart.clear();
-		time = 0;
-		isRunning = true;
-	}
-
-	// 현재 시간 전에 도착한 프로세스들을 processQueue에서 waitingQueue로 이동
-	while (!pQ->empty() && time >= pQ->top().GetArrivalTime()) {
-
-		wQ.Push(pQ->top());
-		pQ->pop();
-	}
-
-	// Dispatch process from waitingQueue to CPU
-	if (!wQ.Empty()) {
-
-		// Preemptive
-		if (isPreemptive) {
-
-			// 
-			//
-			// 
-			//
-			//
-			//
-		}
-
-		// Round-Robin
-		else if (isRoundRobin) {
-
-			//
-			//
-			// 
-			//
-			//
-			//
-		}
-
-		// Non-preemptive and no time-quantum
-		else {
-
-			currentProcess = wQ.Top();
-			time += currentProcess.GetBurstTime();
-			wQ.Pop();
-		}
-	}
-	// If waitingQueue is empty, empty process
-	else {
-
-		currentProcess.SetPid("");
-		time = pQ->top().GetArrivalTime();
-	}
-
-	// ganttChart에 기록
-	ganttChart.emplace_back(currentProcess.GetPid(), time);
-	// scheduling 종료 조건
-	if (pQ->empty() && wQ.Empty()) {
-
-		isRunning = false;
-		pQ = nullptr;
-	}
-}
-
-void CpuScheduler::Reset()
-{
-	isRunning = false;
-	pQ = nullptr;
-	wQ.Clear();
-	ganttChart.clear();
-}
+#endif
