+++ SchedulingSimulator/waiting_queue.h
@@ -1,7 +1,7 @@
-#pragma once
 #include "process.h"
 #include <vector>
-
+#ifndef _WAITING_QUEUE_H
+#define _WAITING_QUEUE_H
 
 
 enum class Scheduling
@@ -46,38 +46,4 @@ private:
 	void Priority(const Process& newProcess);
 };
 
-
-
-void waitingQueue::SetAlgorithm(Scheduling algorithm)
-{
-	this->algorithm = algorithm;
-
-	if (algorithm == Scheduling::FCFS)
-		fp = &waitingQueue::FCFS;
-	else if (algorithm == Scheduling::SJF)
-		fp = &waitingQueue::SJF;
-	else if (algorithm == Scheduling::Priority)
-		fp = &waitingQueue::Priority;
-	else
-		fp = nullptr;
-}
-
-
-
-void waitingQueue::SJF(const Process& newProcess)
-{
-
-}
-
-void waitingQueue::Priority(const Process& newProcess)
-{
-	for (auto iter = data.cbegin(); iter != data.cend(); iter++) {
-
-		if (iter->GetPriority() > newProcess.GetPriority()) {
-
-			data.insert(iter, newProcess);
-			return;
-		}
-	}
-	data.push_back(newProcess);
-}
+#endif
